msmtpq-ng: Fix "Fails with Busybox flock"
With this patch msmtpq-ng (and msmtpq-ng-queue) fail with Busybox flock
due to bb flock not support the -w option, which return an error after
a specified number of seconds of failing to get a lock.
We work around this by using a non-blocking lock attempt (-n) and
counting time ourselves (with some extra effort required due to wanting
to catch child command return codes.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>