tree-wide: Implement per-user queues and lock for system MTA
Closes: LP# 1812795 "System mail send as non-root fails unless member of msmtp group"

This commits improves operation as a system MTA by ensuring that mail
sent by non-root daemons is actually queued and/or sent instead of
failing due to permissions unless made a member of msmtp group.

Shellcheck more scripts (the MTA wrappers) and fix a fatal
quoting error found with it as part of that process.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>