tree-wide: Use separate config from system default
Avoid having msmtpq-ng{-mta} pollute user or system config space for msmtp
by using a separate client configuration for msmqtpq-ng{-mta} only.

Implement a default config for msmtpq-ng{-mta} specific msmtprc file so that
user has less to configure to get mail delivered (in the debian
packaging).

The change to not using the default system-wide or user msmtprc results in
the need to update the test scripts.  Also this exposed problems with the
queuing tests, which this patch fixes.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>